+++ pkg/analysis_server/lib/src/services/correction/assist_internal.dart
@@ -25,6 +25,7 @@ import 'package:analysis_server/src/services/correction/dart/convert_into_is_not
 import 'package:analysis_server/src/services/correction/dart/convert_map_from_iterable_to_for_literal.dart';
 import 'package:analysis_server/src/services/correction/dart/convert_part_of_to_uri.dart';
 import 'package:analysis_server/src/services/correction/dart/convert_quotes.dart';
+import 'package:analysis_server/src/services/correction/dart/convert_to_dot_shorthand.dart';
 import 'package:analysis_server/src/services/correction/dart/convert_to_expression_function_body.dart';
 import 'package:analysis_server/src/services/correction/dart/convert_to_field_parameter.dart';
 import 'package:analysis_server/src/services/correction/dart/convert_to_generic_function_syntax.dart';
@@ -105,6 +106,7 @@ const Set<ProducerGenerator> _builtInGenerators = {
   ConvertMapFromIterableToForLiteral.new,
   ConvertPartOfToUri.new,
   ConvertSwitchExpressionToSwitchStatement.new,
+  ConvertToDotShorthand.new,
   ConvertToDoubleQuotes.new,
   ConvertToExpressionFunctionBody.new,
   ConvertToFieldParameter.new,